requirements.txt
# ===== Flask Core =====
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.5.1
Werkzeug==3.1.3

# ===== Database (Railway Managed MySQL) =====
PyMySQL==1.1.2
cryptography==44.0.0  # Required for secure MySQL auth

# ===== WebSockets / SocketIO Runtime =====
eventlet==0.39.0
gunicorn==23.0.0

# ===== JWT Auth =====
PyJWT==2.10.1

# ===== External APIs =====
openrouteservice==2.3.3
requests==2.32.4

# ===== Utils / ENV =====
python-dotenv==1.2.1

# ===== Scraping (optional but fine) =====
beautifulsoup4==4.13.4
Flask-APScheduler==1.13.1


# ⚠ Removed for size + compatibility:
# torch / torchvision / torchaudio / accelerate
# (Huge wheels not required unless you're running vision/audio AI models)

# If you want AI chatbot inference, use a SMALL model via HF API later
# instead of installing full ML stack in container.
